Ensure GCHP option to write restart via o-server is disabled if < 1000 PEs
Prior to this update the o-server option was turned on if using greater
than 1000 cores but was not turned off again if the number of cores was
subsequently reduced. This update both explicitly turns off the o-server
option if less than 1000 cores, and changes the criteria to turn it on
from >1000 cores to >=1000 cores.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>