Add bug fix for latidue edge calculation in gc_grid_mod.F90
In commit 7f6021c, an error was introduced in the latitude edge calculation
that caused incorrect values in global simulations. The issue was the removal
of an ELSE statement after defining State_Grid%GlobalYEdge for the South
Pole. This is restored here and grid definitions for global 4x5 simulations
are now consistent with those in 13.2.0. This fix resolves Github issue
https://github.com/geoschem/geos-chem/issues/886.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>