Updated the criteria for calling SEASALT_CHEM in sulfate_mod.F90
The call to SEASALT_CHEM (which computes AlkA and AlkC, which then gets
stored as State_Chm%SSAlk(:,:,:,1) and State_Chm%SSAlk(:,:,:,2)) was
not being executed when sulfate chemistry is handled by KPP.  This
resulted in KPP rate-law functions getting zero seasalt alkalinity,
which adversely impacted the computation of rxn rates for e.g. BrSALA.

We have fixed this situation by calling SEASALT_CHEM if:

(1) Sulfate chemistry is handled in sulfate_mod, or
(2) Sulfate chemistry is handled by KPP, and FullRun=FALSE

We will eventually abstract the computation of cloud pH and seasalt
alkalinity out of sulfate_mod.F90, but this commit fixes the issue
for now.

Also updated some comments.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>