Use gammaLiq = gammaIce = 0.01 in function IONO2uptkByH2O
In the original geoschem/geos-chem PR #719, the gammaLiq and gammaIce
values for IONO2 were not specified in routine CloudHet.  This was
a bug in the original development branch.

Tomas Sherwen and Xuan Wang have clarified that the values of
gammaLiq and gammaIce for IONO2 should both be 0.01.  We now pass
these to the CloudHet function in rate law function IONO2uptkByH2O
in KPP/fullchem/fullchem_RateLawFuncs.F90.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>