Fix registering species-based quantities in state_chm_mod.F90
This commit fixes the issue described in geoschem/geoschem #941.

When registering species-based quantities, there was an indexing
error for quantities with less than the total number of species
(i.e. drydep and wetdep quantities).

We have updated the algorithm in the routine Get_Diagnostic_Name
to compute the modelId from the drydep or wetdep ID, and then use that
to look up the name of the species.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>