Rename "RDI" to "RUNDIR" for automated run directory creation updates
Run directory variables were previously prefixed with RDI (Run Directory
Initialization). It was not immediately obvious what RDI stood for, so we
have replaced RDI with the more descriptive RUNDIR. This is a simple swap
where all uses of RDI, rdi, or RDI_ have been replaced with RUNDIR, rundir,
or RUNDIR_. The run directory variables are now saved to a file named
rundir_vars.txt.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>