Bug fix: avoid double-counting of HCl uptake on alkaline sea salt aerosols
This fix corrects the issue first reported by Viral Shah in
geoschem/geos-chem #1034.  We only need to do HCL uptake when
SO2 chemistry is called with when FullRun=.TRUE..  The quick fix was
to add FullRun to the IF block where the computations occur.

NOTE: This issue does not occur in the new KPP sulfur chemistry code:
KPP/fullchem/fullchem_SulfurChemFuncs.F90.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>