In CALC_HEFF, allow negative pKa, but skip missing values
This commit fixes the issue originally raised in geoschem/geos-chem #1001.
Because missing values in the species database are -999, the condition
IF ( pKa > 0 ) in CALC_HEFF (in module GeosUtil/henry_mod.F0) basically
ignores pH in the effective Henry's law constant computations for cloud
water in heterogenous chemistry.

We have now set the condition to IF ( pKa > -100 ), which allows for
negative pKa values, but which will skip missing values.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>