Point to updated Hg restart file in GEOSCHEM_RESTARTS/v2021-12
We have had to update the restart file for the GEOS-Chem Hg simulation
with KPP in order to account for the new species.  Now point to a new
restart file in the ExtData/GEOSCHEM_RESTARTS/v2021-12 folder.  Updated
the download_data.py, download_data.yml, and GC-Classic createRunDir.sh
files accordingly.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>