Add flexible LOG_MSG subroutine in Error_Mod
Both GC-Classic and GCHP can now pass messages to the routine
Log_Msg in Error_Mod. When running in GC-Classic, this simply
prints to stdout. However, when running in GCHP, this utilizes
the pfLogger capability. Two optional arguments allow users to
provide the source of the message and the level of the message
(either debug, info, or warning), allowing more nuanced control
of output.