Input_Opt and input.geos modified for RRTMG FDH
Two options are now present in the RRTMG input menu to handle
the fixed dynamical heating approximation. The user can specify
that they do or do not wish to use the fixed dynamical heating
approximation. If this is set to T (stored as Input_Opt%RRTMG_FDH)
then the user can also set "Read dynamical heating". When set to F,
the code will estimate the dynamical heating rate; this can be
saved through diagnostics as the field DynHeating. When set to T,
the code expects DynHeating to be supplied externally, and uses it
to calculate a temperature difference (using the FDH assumption).
This commit only puts the logical flags in place. It does not
make them functional.