First attempt at RRTMG SEFDH
Users can now select "seasonally evolving fixed dynamical heating".
This allows the stratospheric temperature to evolve over time, as
per Forster et al (1997). To enable this, a new restart variable is
defined (TStrat_Adj) which tracks the stratospheric temperature
adjustment as estimated for the next RRTMG calculated. Using the
SEFDH is meant to provide greater accuracy, and reduces the number
of longwave calculations required; it also allows the (usually
small) effect of temperature adjustment on shortwave forcing to
be included. However, it is estimated to require a ~2 year spinup
before the stratosphere will equilibrate, and (as with the FDH
simulation) requires a reference simulation for comparison.