Kludge for TOMAS: Reset negatives to 1e-26 in DO_TEND (mixing_mod.F90)
In the TOMAS40 simulation, we have noticed that negative values in
routine DO_TEND mixing_mod.F90 trip the error trap.  Betty Croft
@ WashU has added a fix to reset these negatives to 1e-26 in order to
allow these runs to continue.  The origin of the negative values
is still under investigation.

We have added this fix inside an #ifdef TOMAS block so as not to
impact any other simulations.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>