Add option to CH4 simulations for performing analytical inversions
An option to run analytical inversions has been added to the CH4 Menu
in input.geos. This option requires a state vector file be specified in
HEMCO_Config.rc to define the state vector elements, i.e. grid
boxes or clusters of grid boxes in which emissions will be perturbed. The
perturbation value may also be changed in input.geos. The option is set up
so that each state vector element would have its own run directory, and the
output of all perturbation simulations can then be used to construct the
Jacobian. This workflow is automated in the Integrated Methane Inversion
(IMI). Finally, users also have the option to run a posterior simulation
and apply emission and/or OH scale factors derived from an inversion.

References:
  Maasakkers, J.D., D.J. Jacob, M.P. Sulprizio, T. Scarpelli, H. Nesser,
   J.-X. Sheng, Y. Zhang, M. Hersher, A.A. Bloom, K.W. Bowman, J.R.
   Worden, G. Janssens-Maenhout, and R.J. Parker, Global distribution of
   methane emissions, emission trends, and OH concentrations and trends
   inferred from an inversion of GOSAT satellite data for 2010-2015, Atmos.
   Chem. Phys., 19, 7859-7881, 2019.

  Varon, D.J., D.J. Jacob, M. Sulprizio, L.A. Estrada, W.B. Downs, L. Shen,
   S.E. Hancock, H. Nesser, Z. Qu, E. Penn, Z. Chen1, X. Lu3, A. Lorente,
   A. Tewari, and C.A. Randles, Integrated Methane Inversion (IMI 1.0): A
   user-friendly, cloud-based facility for inferring high-resolution
   methane emissions from TROPOMI satellite observations, submitted to
   Geophys. Model Dev., 2022.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>