Now copy the fullchem restart file for the initial TagO3 restart file
In order to ensure that the Tagged O3 simulation starts with the same
global O3 initial conditions as the fullchem simulation, we now copy
the default fullchem restart file to the tagged O3 run directory.

In addition, we have restored changing the HEMCO time cycle flag for
the SPC_RST container from "EFYO" to "EY".  This will bootstrap
all of the other tagged species.

It is recommended that tagged O3 users spin up their simulations
for a sufficient number of years until the O3 is ins steady-state.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>