Bug fix: Set DRYCOEFF = 0 to avoid FPE in HEMCO soil NOx when drydep is off
The DRYCOEFF array is read from disk in Read_DryDep_Inputs (in
GeosCore/drydep_mod.F90).  However, a pointer to DRYCOEFF is also
used in the HEMCO SoilNox extension.

We now set DRYCOEFF to zero in the same line where it is declared (since
it is already a global SAVEd variable), so that it will always be
initialized to zero regardless of whether dry deposition is on or off.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>