Turn off sea salt aerosol debromination and make it an option instead
In this commit we disable sea salt aerosol debromination and include it as
an option instead. This came out of concern over the GEOS-Chem 13.4.0
benchmark where ozone concentrations are too low, especially over the
Southern Ocean.

Xuan Wang wrote:

   I think there may be another much easier way to do this. We can simply
   make [Br-] in aerosols as a very small constant in gckpp_HetRate. I
   remember this is what we do before version 11-2.

Viral Shah wrote:

   I think that would certainly be easier and better. We could set [BrSALA]
   and [BrSALC] to 1e-20 in fullchem_HetStateFuncs/Halide_conc, and that
   would essentially zero-out the rates for all reactions involving BrSALA/C.
   It would also take care of another issue. For Br reactions in clouds, we
   calculate a branching ratio for reactions with HBr vs. BrSALA/C, based on
   the relative amounts of these species. Since we don’t want BrSALA/C to
   participate in the reactions, we'll have to override this calculation and
   set the HBr branching ratio for these reactions to 1. But if [BrSALA/C]
   are small values, then the HBr branching ratio will just be 1.

Here, we add an option to the Chemistry Menu in input.geos to turn sea salt
debromination on/off (default=off, or false). If that option is set to false
then the concentration of BrSALA and BrSALC will be set to 1e-20.

This addresses Github issue #1145.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>