GCHP fix: move setup_wetscav so C_H2O is not all zeros in first timestep
This bug fix removes the differences in H2O2 concentrations introduced
in wet scavenging when breaking up a GCHP run into multiple segments.
Before this fix the call to setup_wetscav was before the first call to
AIRQNT, when State_Met%PMID was still zero. State_Met%C_H2O is calculated
from PMID in setup_wetscav, and thus was all zeros when wet scavenging was
first called in the run.

This fix does not impact GEOS-Chem Classic, but it will introduce small
differences in the GCHP benchmark.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>