Clean up extra GCHP internal state variables
There are now ten internal state arrays beyond species concentrations
and HEMCO restart variables that are on by default for GCHP. Nine of these
were already present before this update. One additional one is now added
to avoid differences in chemistry when splitting up a run. This new
array is AeroH2O_SNA which corresponds to State_Chm%AeroH2O(:,:,:,NDUST+1)
which is computed in isorropia and used in RDAER and full chemistry.

The JNO2 and JOH arrays in the internal state are included for further
testing.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>