Revert to using GLOBAL_Br and GLOBAL_BrO in HEMCO/MERCURY/v2022-04
We realized that the units that we were using for oxidants fields (taken
from the 10 year benchmark) were incorrect.  Those were in [v/v dry[,
but we needed [molec/cm3].  For this reason, we have updated the default
HEMCO_Config.rc.Hg template file and the calls that read HEMCO data
from disk in mercury_mod.F90 to read from the data files in
HEMCO/MERCURY/v2022-04.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>