Prevent div-by-zero in routine MMR_Compute_Flux (in emissions_mod.F90)
This commit fixes the issue reported in geoschem/geos-chem #1215.
We now make sure to skip computation of Flux if Total_Area = 0,
in order to avoid a div-by-zero condition.

We also now use logical variables to indicate if the current species
is the 90-day global, NH, or SH passive species.  This will reduce
the number of string comparisons, and is more efficient.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>