Fix GCHP integration test bug: turn off monthly diagnostics
The default diagnostic settings for benchmark and transport tracers
simulations are to have monthly diagnostics enabled. This makes the
frequency and duration in HISTORY.rc be ignored. The 1-hour integration
tests for these simulations therefore did not have diagnostics written
out. This update modifies the default setting used for those simulations
when creating integration tests to have monthly diagnostics disabled.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>