Fix bug in TotalOC diagnostic that cause differences when breaking up a run
If POA or OPOA are true then TotalOC gets reset every timestep. If neither
are true, and LSOA is true, then TotalOC is increased from its current
value. The problem with this is we do not set State_Diag%TotalOC to zero
prior this computation. This causes TotalOC to be the cumulative value
rather than the instantaneous value if POA and OPOA are false and LSOA is
true. This causes erroneous values, and also makes the values different
depending on how you split up your run into multiple jobs.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>