Use modified EDGAR diurnal scale factors file to correctly read data
The EDGARv42/v2015-02/NO/EDGAR_hourly_NOxScale.nc file uses missing values
and a _FillValue attribute value of 1 to indicate where scale factors are
1.0. This causes a silent bug in GC-Classic where zero values are used
instead since HEMCO treats all missing values as 0 rather than use the
_FillValue.

This also causes a run fail in GCHP after the recent MAPL update if the
_FillValue fix included is enabled. This is because data with values
equivalent to _FillValue are now treated as missing and not included in
regridding. For this reason that fix is currently disabled MAPL. We will
enable it following this update.

This commit updates the GC-Classic and GCHP config files to use
EDGARv42/v2015-02/NO/EDGAR_hourly_NOxScale.modified.nc, which is a
duplicate of the non-modified file but with all missing data removed and
replaced with value 1.0.

This update causes small differences in GC-Classic. No differences are
expected for GCHP.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>