Update OFFLINE_BIOVOC and OFFLINE_SOILNOX directories
The entries in HEMCO_Config.rc and ExtData.rc have been updated to use
the v2021-12 directories for the OFFLINE_BIOVOC and OFFLINE_SEASALT
emissions.

Hongjian Weng wrote:

   We have uploaded the new version OFFLINE_BIOVOC and OFFLINE_SOILNOX at
   native MERRA-2 (0.5 x 0.625, from 1980 to 202012) and GEOS-FP (0.25 x
   0.3125, from 2014 to 202009).

   These emissions were generated with GEOS-Chem 13.2.0 and the updates
   include:

   1. Use new version Yuan-processed MODIS LAI (v2021-06);
   2. Fix the unit bug for soil NOx emission;
   3. Add methanol(MOH) variables for biogenic VOCs emissions;
   4. All emissions are now all in kg species/m2/s instead of kgN/m2/s or kgC/m2/s before.

   For more information, please see README files, which I have also
   uploaded to Washu.

In this commit, we also change the time cycle flag in HEMCO_Config.rc from
"C" to "EF" to ensure users are aware when running outside of the available
data range (because HEMCO will crash with a file-not-found error now). In
this case users can opt to use the online emissions, latest available year,
or generate a climatology.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>