Comment out budget diags and remove most @s in GCHP HISTORY.rc template
This update is only for the fullchem HISTORY.rc template. Presence of
@ symbols prefacing lines in the template result in commenting out those
lines when creating a benchmark simulation. This was causing a problem
for the 10-year benchmark which uses more diagnostic collections than
the 1-month and 1-year benchmarks. The @s are now removed from collection
definitions, although they remain for the collection toggles at the top
of the file.

To avoid a future file-too-long error in ESMF most of the budget diagnostic
fields are now commented out, with a new message at the top of the file
about checking collection fields to make sure the ones you wish to output
are uncommented. Commented out lines are not included in the character
count when ESMF reads the file as one string. Note that the current
HISTORY.rc is not too long to cause this error even when all collection
definitions are uncommented. However, I commented out nearly all the
budget diagnostics just in case, and because the budget diagnostics are
rarely all needed in their entirety. Users should uncomment only the ones
they want to output.

Fyi, the possibility of a file-too-long error in ESMF will no longer be
there after we update to MAPL 3 and switch to using yml files.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>