Revert GCClassic_Output/14.0.0 data to GCClassic_Output/13.0.0
The GCClassic_Output/14.0.0 fields that were pushed to the HEMCO data
folder were created with a version of GEOS-Chem that had an error in
the pressure-fixer.  This had caused much higher differences in long-
lived species such as CO when compared to GC versions prior to the fix.

Until the GCClassic 10-year benchmarks can be re-run, we will revert
the GCClassic_Output/14.0.0 data paths to GCClassic_Output/13.0.0
in the run/shared_settings/gmao_hemco.txt file.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>