Clean up and documentation of GCHP runScriptSamples/operational_examples
This update removes the adjoint run script because it has no contact
information and appears to be the same as the NASA Pleiades script.
The contact information for NASA Pleiades is now WashU contingent of
GCST.

Files in the operation_examples folder should be expanded and edited
as needed by users of specific clusters, and may be pulled into any
GEOS-Chem version on a rolling basis.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>