Set species concentration arrays as pointers to internal state in GCHP
This commit removes allocation of State_Chm%Species(N)%Conc arrays to
avoid duplication in memory when using GCHP. Species concentration arrays
are now stored only in the MAPL internal state. Because the internal state
level convention is top-of-atmosphere=1, the species values in the
internal state are vertically flipped at the start and end of the
GEOS-Chem Run_ method. For safety, the GEOS-Chem species concentration
array pointers are only associated during this time.

There are some modifications in this commit for the same update in GEOS
but they are not yet complete.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>