Remove GCHP placeholder code for GEOS provider services
The GEOS provider code was originally abstracted a while back in
preparation for moving it out of Chem_GridCompMod.F90. That work is now
done within new files in Interfaces/GEOS. The original code can now be
safely deleted. Despite being intended for use with GEOS, it was all
in GCHP-only blocks for testing purposes.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>