Update CESM/HISTORY.rc.
Because we will be adopting the GCHP HistoryExports component for
CESM-GC, we need to write out the wildcards explicitly in HISTORY.rc
similarly to GCHP.

Because not all processes in CESM-GC are handled within GEOS-Chem,
the collections of processes not handled by GEOS-Chem routines,
and as such diagnostics would be unavailable, are deleted from HISTORY.rc
in CESM to reduce confusion. This includes drydep, wetdep, advection.