Update EDGARv6 files for CH4 simulations to fix timestamp bug
As reported in https://github.com/geoschem/geos-chem/issues/1484, the
files in HEMCO/CH4/v2022-01/EDGARv6 had the incorrect timestamps which
caused data for the wrong date to be used in CH4 simulations. The files
have been reprocessed and the corrected files may be found in
HEMCO/CH4/v2022-11/EDGARv6. Here we update the filepaths in HEMCO_Config.rc
for CH4 and tagCH4 simualtions to use the corrected files.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>