Add fixes to restore the analytical inversion option for CH4 simulations
The conversion from input.geos to geoschem_config.yml broke several of
the switches for the CH4 analytical inversion options. Yo fix this,
"ch4_simulation_options%" needed to be added to the key passed to QFYAML
in order to read the correct value and pass it back to Input_Opt.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>