Preliminary patch for removal of module variables from ucx_mod.F90.
This patch moves in-module variables in UCX_MOD to State_Chm, to facilitate integration of GEOS-Chem into the WRF-GC and CESM models.

There are some shortcomings with this patch, namely because the UCX variables are not regularly-sized (not I, J, L), they cannot be allocated using the current implementation of Init_and_Register, and instead are allocated by ucx_mod itself. This may need an alternative implementation of State_Chm_Mod's Init_and_Register to fix.