Revert updates to request partition in GCHP integration test scripts
In commit f2ce7d44 updates were made to allow for specification of SLURM
partition for GCHP integration tests. However, this didn't work properly
because gchp.slurm.sh script was not located in the top level directory.
For now we will remove this capability. They may come in later with PR
https://github.com/geoschem/geos-chem/pull/1565.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>