HEMCO bug workaround: configure GFAS_pFe as second read of GFAS_SO2 input
This update is a work-around for a problem interpreting the dash symbols
in HEMCO_Config.rc for the case of GFAS_pFe when using vertical injection
levels (setting xyL=1:scal300). It replaces the dashes with input
information for GFAS SO2. The dashes are supposed to make this automatic.
This is a temporary fix until the HEMCO bug can be fixed.

This update also includes ExtData.rc entries for GFAS. Previously it was
left to the user to add these if turning on GFAS. The entries are
commented out by default.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>