Bug fix to only write species metadata yaml if file specified in config
Previously the species metadata yaml file write subroutine was always
executed since the criteria to skip it was if Input_Opt%SpcMetaDataOutFile
was not equal to "none". Since the field was initialized as an empty
string this would return true and writing the file would be attempted.
This results in model crash if a file path is not present in config file
geoschem_config.yml. This is fixed by changing the criteria to check
string length rather than value. The yaml file will only get written if
an output path is specified in config file geoschem_config.yml.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>