Fix bug in RRTMG Radiation diagnostic indexing
This bug impacts only radiation AOD, SSA, and Asym diagnostics.

RRTMG State_Diag array indexing is different than RRTMG array
indexing since the State_Diag arrays are only allocated for the
diagnostics that are enabled in HISTORY.rc. For example, if no PM
diagnostics are enabled in HISTORY.rc then the diagnostic arrays will
be reduced in size to not include PM.

This bug still used the RRTMG indexing for a few of the State_Diag
arrays, causing an out of bounds error if a reduced set of diagnostics
was in HISTORY.rc, and writing incorrect data since the ordering of the
two arrays are different.


Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>