Remove X-HRS printout from GEOS-Chem log file
Each time the date and time is printed in the GEOS-Chem log file, X-HRS
is also printed. For example:

   ---> DATE: 2019/07/01  UTC: 00:00  X-HRS:      0.000000
        - DO_LINEAR_CHEM: Linearized chemistry at 2019/07/01 00:00
        - LINOZ_CHEM3: Doing LINOZ
   ---> DATE: 2019/07/01  UTC: 00:10  X-HRS:      0.166667
   ---> DATE: 2019/07/01  UTC: 00:20  X-HRS:      0.333333

X-HRS represents the hours since the start of the run (as defined in
GeosUtil/time_mod.F90). It is removed here since the name and format of
this variable can be confusing and isn't needed in the log file.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>