Require GCHP restart files to contain all species for some simulation types
This update is limited to GCHP only and does not impact GEOS. It
replicates behavior already in GC-Classic. There is a new switch in
GCHP.rc to change whether species are required in the restart file. This
is controlled from setCommonRunSettings.sh for convenience. All integration
tests will allow missing species regardless of simulation type.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>