Add RxnConst diagnostic for archiving reaction rate constants.
This pull request implements the RxnConst_?RXN? diagnostic, allowing for archival of reaction rate constants (units vary) from KPP.

It is not enabled by default but can be enabled in a similar manner to the existing RxnRate_ diagnostics, which output the reaction rates (in molec cm-3 s-1) which are the rate constants with the species concentrations already applied.

Minor cosmetic fixes to the comments and metadata are added to describe the correct units of the RxnRate diagnostic (molec cm-3 s-1) as well.

Signed-off-by: Haipeng Lin <hplin@seas.harvard.edu>