Update GFED4 climatology file to use 2010-2019 range
The GFED4 climatology option now points to the GFED4_climatology_2010-2019_05x0625.nc
file instead of GFED4_climatology_2007-2021_05x0625.nc. The 2007-2021 file
still remains in the HEMCO/GFED4/v2023-03 option, as does the file containing
monthly emissions for 2007-2021 in case users want to compute climatologies
for other year ranges. The 2010-2019 range is selected for consistency with
the other climatology options used in GEOS-Chem and is the same year range
as the 10-year benchmark simulation.

Also updated CHANGELOG.md.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>