Use first timestep current time rather than ESMF clock start for GC start
The ESMF Clock start time in MAPL is set to the entry for BEG_DATE in
config file CAP.rc. This start time is not the same as the simulation
start if config file cap_restart is present as the start of the run. Using
cap_restart is now default in GCHP. This update fixes a bug where
GEOS-Chem time_mod.F90 start time was BEG_DATE rather than simulation
start.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>