Add fix for copying CH4 emissions when it's not a tagCH4 simulation
When copying CH4_EMIS into eflx in hco_intergace_gc_mod.F90, it was
automatically assumed that if NA (number of advected species) > 1 then
it was a tagged CH4 simulation and to copy the sector-specific emissions.
However, we want to allow for flexibility to run with more than one CH4
tracer in a non-tagCH4 simulation so we specifically split off the code
for CH4 and tagCH4 simulations with IF statements.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>