Fix duplicate entries for O3_Loss in HEMCO_Config.rc.TransportTracers for GCHP
The HEMCO_Config.rc.TransportTracers file included both the {RUNDIR_O3_LOSS}
token and the actual entry for O3_LOSS which resulted in the entry getting
listed twice in the HEMCO_Config.rc that ships with run directories. The
line has been fixes so that just the token is used in the template file.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>