Rename GCHP mid-run restart files in run scripts even if run crashes
This update moves the renaming of mid-run restart files to after the
renaming of the end-of-run restart file. This prevents mid-run checkpoint
files from having non-standard names if the run crashes at some point.
Previously the renaming of mid-run checkpoint files did not happen if the
run crashed because the run script would exit prior to reaching execution
of that code.

This update also include deletion of the end-of-run restart file
Restarts/gcchem_internal_checkpoint if the run fails. Depending on where
in the run the failure occurs this file may exist. Its presence will
prevent a future run and so it must be deleted.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>