Use standard format for AEIC 2019 monthly climatolog in ExtData.rc
The standard format used for monthly climatology in ExtData.rc sets
the clim flag to Y and the read frequency to F{year}-%m2-01T00:00:00.
The previous specification using clim year for the climatology flag and
including tokens for year, month, and day for read frequency works but
it would be good to use the standard format for consistency and to avoid
confusion during the migration from the rc file to the yaml file.

This is a zero diff update for GCHP since the old and new formats give
identical results.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>