Adjust relative tolerance value if first KPP integration attempt fails
Following the chemistry updates in 14.2.0, users are encountering KPP
convergence errors more frequently. To fix this, RTOL was changed from
0.5e-2 to 0.5e-3, but this drastically slowed down simulations so it was
changed back to 0.5e-2. Because we are still seeing these convergence
errors, one workaround is to adjust the RTOL value to 0.5e-3 after the KPP
integration fails the first time. If it fails twice, GEOS-Chem will still
crash as usual. The RTOL value resets to 0.5e-2 for the next grid box in
the loop.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>