Add GCHP config file for ESMF logging; turn off ESMF logging by default
This update has an accompanying update in MAPL to read config file ESMF.rc
to retrieve the setting for ESMF_LogKind. Previously this was set at
compilation time in GCHP and was hard-coded to write an ESMF log file
for each thread upon error. There is currently a benign ESMF error
triggered sometimes when monthly diagnostics are enabled, and that is the
motivation for turning off ESMF logging by default, at least temporarily.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>