Remove double-counting of IONO2 loss (cf geoschem/geos-chem #1880)
In geoschem/geos-chem #1880, @viral211 suggested that we are
double-counting the loss of IONO2 by alkaline SALA and SALC in
hetchem routines:

- IuptkByAlkSALA1stOrd
- IuptkByAlkSALC1stOrd

and then again in routine

- IONO2uptkByH2O

To remove the double-counting, we have removed the code that
handles the removal of IONO2 by alkaline sea salt aerosol.
A more permanent solution by @beckyalexander may be forthcoming
in the near future.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>