Revert PR #1702: Restore consideration of sea salt alkalinity to heterogenerous acid-catalyzed reactions of halogens on sea salt aerosols
Pull request #1702 fixed bugs in heterogeneous chemistry by restoring sea
salt alkalinity to certain reactions in the halogen chemistry. That update
introduced many KPP integrate error messages into the log file. The errors
were overlooked because the integration succeeded on the second attempt.
In the 1-year full-chemistry benchmark for 14.2.0-rc.0, a sharp decrease in
O3 was observed at several locations in the model vs observation plots. It
turns out that decrease could be attributed to the changes in #1702. As a
result, we now revert those updates to restore model performance. Implementing
these fixes at a later date is a science issue and should be fully validated.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>