Assorted GCHP rundir updates related to mass flux/winds in GEOS-FP/IT
Since there are multiple config settings for GEOS-FP and GEOS-IT the
shared/settings folders for those met fields are now in subfolders
geosfp and geosit. The folder geosit has a subdirectory called discover
for settings files specific to the discover cluster which uses different
GEOS-IT filenames and directory structure.

The GCHP run directory names are also changes to put met after
simulation type. This allows a more extended met description for the
case of using native fields and whether you are using mass fluxes,
winds, or derived winds.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>