Set RUNDIR_GRID_HALF_POLAR for global 0.25x0.3125 and 0.5x0.625 grids in createRunDir.sh
Previously, when a user selected the global domain for the 0.25x0.3125 or
0.5x0.625 resolutions, the $RUNDIR_GRID_HALF_POLAR option was not defined, so
the corresponding token in geoschem_config.rc was not replaced. This is now
fixed.

Note: Typically users don't select the entire globe when using these resolutions
but selecting these options when setting up a run directory may assist in
defining custom regional domains.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>