Add option to perturb OH in CH4 simulations
A new option `OH_perturbation_factor` has been added to geoschem_config.rc
for CH4, tag CH4, and carbon simulations. This option is added to facilitate
optimizing for OH in the Integrated Methane Inversion (IMI). The scaling of
OH is currently done in global_ch4_mod.F90, but it would be more straightforward
to do this in HEMCO_Config.rc.

Also in this commit, other CH4 options have been slightly renamed for
clarity:
  - Input_Opt%AnalyticalInv -> Input_Opt%DoAnalyticalInv
  - Input_Opt%PerturbEmis -> Input_Opt%EmisPerturbFactor
  - Input_Opt%PerturbCH4BoundaryConditions -> DoPerturbCH4BoundaryConditions

The options for using emissions and OH scale factors from a previous
inversion have been moved to the end of the CH4 options list so that the
options for the Jacobian runs within the IMI framework are grouped together.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>