Move GCHP import OCEAN_MASK used in transport tracers to met input templates
This update fixes a bug where OCEAN_MASK was incorrectly defined when using
native GEOS-FP input data. The template in ExtData.rc.TransportTracers
assumed processed input data which is an over-generalization. Defining
the ocean mask input is now moved to where all other met-field inputs are
defined for GCHP, in the met-specific files in run/shared/settings/.

NOTE: OCEAN_MASK is a duplicate of input FROCEAN. However, it is needed
since HEMCO does not draw upon existing met-field FROCEAN for the ocean
masking.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>