Update sample restart files for fullchem and TransportTracers to 14.2.0 benchmark files
Out-of-the-box fullchem and TransportTracers simulations for 14.2.0 and 14.2.1
are failing because additional species were added in 14.2.0. The time cycle flag
used in HEMCO_Config.rc (EYFO) forces GEOS-Chem to crash with an error when
species are not found in the restart file. To avoid this, we need to update
the sample restart files for these simulations to restart files saved out from
the 14.2.0 1-year benchmark simulations.

Closes https://github.com/geoschem/geos-chem/issues/1986

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>